Add files via upload
This tool asks the user how many scores will be graded. valid scores range 0-100. if input is invalid, user must enter a valid input to continue to the next input.

After input is checked for validity, it is stored in a list to be used later.
upon completion of input, "results" summary is output. the lowest score is listed, then dropped from the list. The modified list is shown, as well as the modified average score, and a letter grade.